Pre-download transaction for zkevm-demo (#1945)
