+++ examples/zkevm-demo/core/Cargo.toml
@@ -12,6 +12,7 @@ revm = { version = "3.5", default-features = false, features = [
   "serde",
 ] }
 serde = "1.0"
+serde_json = "1.0"
 tokio = { version = "1", features = [
   "rt-multi-thread",
   "macros",
