+++ examples/zkevm-demo/core/testdata/tx.json
@@ -1,0 +1,1 @@
+{"hash":"0x671a3b40ecb7d51b209e68392df2d38c098aae03febd3a88be0f1fa77725bbd7","nonce":"0x70","blockHash":"0x3f841e7f8e56223202e174a94524e33cb7aa3a0cc5141b6efd24be3520655ec7","blockNumber":"0xfa9cc2","transactionIndex":"0xa6","from":"0xc04b70290e1f97ed289da6c22be77526ab1e7493","to":"0x7d1afa7b718fb893db30a3abc0cfc608aacfebb0","value":"0x0","gasPrice":"0x3cb654825","gas":"0xbdc0","input":"0xa9059cbb000000000000000000000000c04b70290e1f97ed289da6c22be77526ab1e749300000000000000000000000000000000000000000000000d8d726b7177a80000","v":"0x0","r":"0x1b9c6857a687afeebcc703ff36269c4ca0f6e75eef07d276e753edb41549f95a","s":"0x5d52121322935c6e17b1c57b8923709ce35bf6bd00f61ce37aa05e159e61da31","type":"0x2","accessList":[],"maxPriorityFeePerGas":"0x9d6c81c","maxFeePerGas":"0x474019874","chainId":"0x1","yParity":"0x0"}
