+++ examples/zkevm-demo/core/testdata/zkdb.json
@@ -1,0 +1,1 @@
+{"basic":{"idx":0,"elms":[{"balance":"0xcab67ebc86b4af","nonce":112,"code_hash":"0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470","code":{"bytecode":"0x","state":"Raw"}},{"balance":"0x270fb525defe3d5a6a0","nonce":0,"code_hash":"0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470","code":{"bytecode":"0x","state":"Raw"}},{"balance":"0x0","nonce":1,"code_hash":"0x8b0a83a4cd51e97515c5898d296297d96c5377467369ec0ff432948c3ae995f2","code":{"bytecode":"0x608060405234801561001057600080fd5b5060043610610128576000357c0100000000000000000000000000000000000000000000000000000000900480635c975abb116100bf5780638456cb591161008e5780638456cb591461031057806395d89b4114610318578063a457c2d714610320578063a9059cbb1461034c578063dd62ed3e1461037857610128565b80635c975abb146102b45780636ef8d66d146102bc57806370a08231146102c457806382dc1ec4146102ea57610128565b8063313ce567116100fb578063313ce5671461023a57806339509351146102585780633f4ba83a1461028457806346fbf68e1461028e57610128565b806306fdde031461012d578063095ea7b3146101aa57806318160ddd146101ea57806323b872dd14610204575b600080fd5b6101356103a6565b6040805160208082528351818301528351919283929083019185019080838360005b8381101561016f578181015183820152602001610157565b50505050905090810190601f16801561019c5780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b6101d6600480360360408110156101c057600080fd5b50600160a060020a03813516906020013561043c565b604080519115158252519081900360200190f35b6101f2610460565b60408051918252519081900360200190f35b6101d66004803603606081101561021a57600080fd5b50600160a060020a03813581169160208101359091169060400135610466565b61024261048c565b6040805160ff9092168252519081900360200190f35b6101d66004803603604081101561026e57600080fd5b50600160a060020a038135169060200135610495565b61028c6104b2565b005b6101d6600480360360208110156102a457600080fd5b5035600160a060020a0316610516565b6101d661052f565b61028c610538565b6101f2600480360360208110156102da57600080fd5b5035600160a060020a0316610543565b61028c6004803603602081101561030057600080fd5b5035600160a060020a031661055e565b61028c61057e565b6101356105e4565b6101d66004803603604081101561033657600080fd5b50600160a060020a038135169060200135610645565b6101d66004803603604081101561036257600080fd5b50600160a060020a038135169060200135610662565b6101f26004803603604081101561038e57600080fd5b50600160a060020a038135811691602001351661067f565b60058054604080516020601f60026000196101006001881615020190951694909404938401819004810282018101909252828152606093909290918301828280156104325780601f1061040757610100808354040283529160200191610432565b820191906000526020600020905b81548152906001019060200180831161041557829003601f168201915b5050505050905090565b60045460009060ff161561044f57600080fd5b61045983836106aa565b9392505050565b60025490565b60045460009060ff161561047957600080fd5b610484848484610728565b949350505050565b60075460ff1690565b60045460009060ff16156104a857600080fd5b61045983836107f1565b6104bb33610516565b15156104c657600080fd5b60045460ff1615156104d757600080fd5b6004805460ff191690556040805133815290517f5db9ee0a495bf2e6ff9c91a7834c1ba4fdd244a5e8aa4e537bd38aeae4b073aa9181900360200190a1565b600061052960038363ffffffff6108a116565b92915050565b60045460ff1690565b610541336108d8565b565b600160a060020a031660009081526020819052604090205490565b61056733610516565b151561057257600080fd5b61057b81610920565b50565b61058733610516565b151561059257600080fd5b60045460ff16156105a257600080fd5b6004805460ff191660011790556040805133815290517f62e78cea01bee320cd4e420270b5ea74000d11b0c9f74754ebdbfc544b05a2589181900360200190a1565b60068054604080516020601f60026000196101006001881615020190951694909404938401819004810282018101909252828152606093909290918301828280156104325780601f1061040757610100808354040283529160200191610432565b60045460009060ff161561065857600080fd5b6104598383610968565b60045460009060ff161561067557600080fd5b61045983836109b3565b600160a060020a03918216600090815260016020908152604080832093909416825291909152205490565b6000600160a060020a03831615156106c157600080fd5b336000818152600160209081526040808320600160a060020a03881680855290835292819020869055805186815290519293927f8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925929181900390910190a350600192915050565b600160a060020a038316600090815260016020908152604080832033845290915281205461075c908363ffffffff6109c916565b600160a060020a038516600090815260016020908152604080832033845290915290205561078b8484846109de565b600160a060020a0384166000818152600160209081526040808320338085529083529281902054815190815290519293927f8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925929181900390910190a35060019392505050565b6000600160a060020a038316151561080857600080fd5b336000908152600160209081526040808320600160a060020a038716845290915290205461083c908363ffffffff610aab16565b336000818152600160209081526040808320600160a060020a0389168085529083529281902085905580519485525191937f8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925929081900390910190a350600192915050565b6000600160a060020a03821615156108b857600080fd5b50600160a060020a03166000908152602091909152604090205460ff1690565b6108e960038263ffffffff610abd16565b604051600160a060020a038216907fcd265ebaf09df2871cc7bd4133404a235ba12eff2041bb89d9c714a2621c7c7e90600090a250565b61093160038263ffffffff610b0916565b604051600160a060020a038216907f6719d08c1888103bea251a4ed56406bd0c3e69723c8a1686e017e7bbe159b6f890600090a250565b6000600160a060020a038316151561097f57600080fd5b336000908152600160209081526040808320600160a060020a038716845290915290205461083c908363ffffffff6109c916565b60006109c03384846109de565b50600192915050565b6000828211156109d857600080fd5b50900390565b600160a060020a03821615156109f357600080fd5b600160a060020a038316600090815260208190526040902054610a1c908263ffffffff6109c916565b600160a060020a038085166000908152602081905260408082209390935590841681522054610a51908263ffffffff610aab16565b600160a060020a038084166000818152602081815260409182902094909455805185815290519193928716927fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef92918290030190a3505050565b60008282018381101561045957600080fd5b600160a060020a0381161515610ad257600080fd5b610adc82826108a1565b1515610ae757600080fd5b600160a060020a0316600090815260209190915260409020805460ff19169055565b600160a060020a0381161515610b1e57600080fd5b610b2882826108a1565b15610b3257600080fd5b600160a060020a0316600090815260209190915260409020805460ff1916600117905556fea165627a7a72305820cce49a7abfc4bd1fcccf59df283bbf2975819bc0cd4e6f6755f46ea9411bdc650029","state":"Raw"}}]},"code_hash":{"idx":0,"elms":[]},"storage":{"idx":0,"elms":["0x0","0x1dc32bb47ac9553a3d"]},"block":{"idx":0,"elms":[]}}
